# internal/interface/http/openapi.yaml
openapi: 3.0.0
info:
  title: Kkachi Server API
  version: 1.0.0
  description: API for Kkachi, a central documentation coordination system.
paths:
  /healthz:
    get:
      summary: Health check endpoint
      description: Returns server health status for deployment monitoring
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
  /api/state:
    get:
      summary: Get server state (Web UI alias)
      description: Alias for /state endpoint, used by the v2 Web UI dashboard
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /docs/head:
    get:
      summary: Get current HEAD commit hash of docs repository
      parameters:
        - name: project
          in: query
          required: true
          schema:
            type: string
          description: Project name
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  head:
                    type: string
                    description: Commit hash
        '400':
          description: Bad Request (missing project or unknown project)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /docs/snapshot:
    get:
      summary: Get snapshot of docs directory
      parameters:
        - name: project
          in: query
          required: true
          schema:
            type: string
          description: Project name
        - name: commit
          in: query
          required: false
          schema:
            type: string
          description: Commit hash (optional, defaults to HEAD)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  commit:
                    type: string
                  snapshot:
                    type: string
                    description: Base64 encoded tar.gz
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /workspaces/register:
    post:
      summary: Register or update a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - project
                - local_path
                - repo_url
                - actor_email
              properties:
                project:
                  type: string
                local_path:
                  type: string
                repo_url:
                  type: string
                actor_email:
                  type: string
      responses:
        '200':
          description: Successful registration
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspace_id:
                    type: string
                  current_docs_head:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: internal_server_error
  /docs/push:
    post:
      summary: Push docs snapshot to repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workspace_id
                - base_docs_hash
                - docs_snapshot
                - actor_email
              properties:
                workspace_id:
                  type: string
                base_docs_hash:
                  type: string
                docs_snapshot:
                  type: string
                  description: Base64-encoded tar.gz of docs directory
                actor_email:
                  type: string
      responses:
        '200':
          description: Push result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  status:
                    type: string
                    enum: [updated, nochange, outdated]
                  new_docs_hash:
                    type: string
                  current_docs_hash:
                    type: string
                  error:
                    type: string
        '400':
          description: Bad Request (missing fields, unknown workspace, or unknown commit)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  error:
                    type: string
        '409':
          description: Conflict (docs repo busy)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  error:
                    type: string
                    example: docs_repo_busy
  /state:
    get:
      summary: Get server state (debug/monitoring)
      description: Returns project docs heads and registered workspace summaries
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  docs_heads:
                    type: object
                    additionalProperties:
                      type: string
                    description: Map of project name to current docs HEAD commit hash
                  workspaces:
                    type: array
                    items:
                      type: object
                      properties:
                        workspace_id:
                          type: string
                        project:
                          type: string
                        docs_repo_id:
                          type: string
                        local_path:
                          type: string
                        repo_url:
                          type: string
                        docs_hash:
                          type: string
                        last_reported_at:
                          type: string
                          format: date-time
                        last_actor_email:
                          type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /projects/{project}:
    delete:
      summary: Delete a project and its docs repo mapping
      parameters:
        - name: project
          in: path
          required: true
          schema:
            type: string
          description: Project name to delete
        - name: force
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Force delete even if workspaces are registered
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: unknown_project
        '409':
          description: Conflict - project has registered workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: project_has_workspaces
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /projects:
    post:
      summary: Add or update a project to docs repo mapping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - project
                - docs_repo_id
                - docs_repo_url
                - actor_email
              properties:
                project:
                  type: string
                  description: Project name
                docs_repo_id:
                  type: string
                  description: Docs repository identifier
                docs_repo_url:
                  type: string
                  description: Git URL of the docs repository
                actor_email:
                  type: string
                  description: Email of the user performing the action
      responses:
        '200':
          description: Project added/updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '400':
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /api/pty/sessions:
    post:
      summary: Create a new PTY session (v3)
      description: Creates a PTY session in the specified workspace directory. Requires Bearer token if auth is enabled.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workspace_id
              properties:
                workspace_id:
                  type: string
                  description: The workspace ID to create the session in
                cwd_rel:
                  type: string
                  description: Relative path within workspace (optional, defaults to workspace root)
                shell:
                  type: string
                  description: Shell to use (optional, defaults to /bin/sh)
                cols:
                  type: integer
                  description: Terminal columns (optional, defaults to 80)
                  minimum: 20
                  maximum: 400
                rows:
                  type: integer
                  description: Terminal rows (optional, defaults to 24)
                  minimum: 5
                  maximum: 200
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                  ws_url:
                    type: string
                    description: WebSocket URL for attaching to the session
                  resolved_cwd:
                    type: string
                    description: Resolved absolute path of the CWD
        '400':
          description: Bad Request (invalid input, unknown workspace, path traversal)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PTYErrorResponse'
        '401':
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PTYErrorResponse'
        '429':
          description: Too Many Requests (session limit exceeded)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PTYErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PTYErrorResponse'

  /api/pty/sessions/{id}:
    delete:
      summary: Terminate a PTY session (v3)
      description: Terminates the PTY session. Idempotent - returns success even if session doesn't exist. Requires Bearer token if auth is enabled.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Session ID to terminate
      responses:
        '200':
          description: Session terminated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PTYErrorResponse'

  /api/pty/sessions/{id}/ws:
    get:
      summary: Attach to a PTY session via WebSocket (v3)
      description: |
        Establishes a WebSocket connection to stream PTY input/output.
        If auth is enabled, requires an `auth_token` cookie.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Session ID to attach to
      responses:
        '101':
          description: WebSocket protocol handshake successful
        '401':
          description: Unauthorized (missing or invalid cookie)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PTYErrorResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PTYErrorResponse'
        '409':
          description: Conflict (session already attached)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PTYErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    StateResponse:
      type: object
      properties:
        docs_heads:
          type: object
          additionalProperties:
            type: string
          description: Map of project name to current docs HEAD commit hash
        workspaces:
          type: array
          items:
            type: object
            properties:
              workspace_id:
                type: string
              project:
                type: string
              docs_repo_id:
                type: string
              local_path:
                type: string
              repo_url:
                type: string
              docs_hash:
                type: string
              last_reported_at:
                type: string
                format: date-time
              last_actor_email:
                type: string
    PTYErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: |
            Error code. Possible values:
            - `invalid_request_body`
            - `missing_workspace_id`
            - `unknown_workspace`
            - `cwd_traversal_attempt`
            - `absolute_path_not_allowed`
            - `invalid_cwd`
            - `shell_not_allowed`
            - `invalid_terminal_size`
            - `session_limit_exceeded`
            - `pty_spawn_failed`
            - `missing_session_id`
            - `session_not_found`
            - `session_already_attached`
            - `unauthorized`
            - `internal_server_error`
        message:
          type: string
          description: Human-readable error message